feat : User-Device구조 -> User-RootUser구조 변경
